schema: add selftest.sh script and fixes for schema errors discovered by it
The selftest.sh script will generate both an example input document and
reader code for the current pretty schema. It will then attempt to parse
the example document using the reader class.

This uncovered a number of inconsistencies in the current schema which
are fixed in this commit as well.

Additionally remove redundant examples for simple types.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>